flask-backend/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
acryl-datahub==0.14.1.12
flask_cors==5.0.0